Update documentation and event handler
