+++ ai4rag/utils/event_handler/__init__.py
@@ -2,4 +2,5 @@
 # Copyright IBM Corp. 2025-2026
 # SPDX-License-Identifier: Apache-2.0
 # -----------------------------------------------------------------------------
-from .event_handler import BaseEventHandler, LogLevel
+from .base_event_handler import BaseEventHandler, LogLevel
+from .event_handler import LocalEventHandler
