--- docs/.pages
@@ -1,2 +1,0 @@
-# This file is intentionally empty
-# Navigation is controlled by mkdocs.yml
