+++ docs/about/changelog.md
@@ -6,8 +6,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [0.1.1](https://github.com/IBM/ai4rag/releases/tag/v0.1.1)
 
-## [0.1.0]
+### Added 
+- Support for `LocalEventHandler`
+- Support for external models introduced via `OpenAI` client
+- CI/CD tooling
+
+
+### Fixed
+- Documentation and `README.md` update
+- Updated samples
+- Updated docstrings
+
+### Changed
+- Loose required parameters for `AI4RAGExperiment`
+
+
+## [0.1.0](https://github.com/IBM/ai4rag/releases/tag/v0.1.0)
 
 ### Added
 - Initial working implementation of `ai4rag` that can be used with `llama-stack` for RAG Template optimization
