+++ docs/architecture/overview.md
@@ -213,60 +213,6 @@ HPO Optimizer (feedback for next configuration)
 
 ---
 
-## Key Abstractions
-
-### BaseFoundationModel
-
-Abstract interface for LLM integration:
-
-```python
-class BaseFoundationModel(ABC):
-    @abstractmethod
-    def generate(self, prompt: str) -> str:
-        """Generate text from a prompt."""
-        pass
-```
-
-Implementations:
-- `LSFoundationModel`: Llama Stack integration
-
-### BaseEmbeddingModel
-
-Abstract interface for embedding generation:
-
-```python
-class BaseEmbeddingModel(ABC):
-    @abstractmethod
-    def embed(self, texts: List[str]) -> List[List[float]]:
-        """Generate embeddings for texts."""
-        pass
-```
-
-Implementations:
-- `LSEmbeddingModel`: Llama Stack integration
-
-### BaseVectorStore
-
-Abstract interface for vector storage:
-
-```python
-class BaseVectorStore(ABC):
-    @abstractmethod
-    def add_documents(self, documents: List[Document]) -> None:
-        """Add documents to the store."""
-        pass
-
-    @abstractmethod
-    def similarity_search(self, query: str, k: int) -> List[Document]:
-        """Search for similar documents."""
-        pass
-```
-
-Implementations:
-- `LSMilvusVectorStore`: Milvus via Llama Stack
-- `ChromaVectorStore`: ChromaDB in-memory
-
----
 
 ## Extension Points
