+++ docs/getting-started/installation.md
@@ -3,8 +3,15 @@
 ## Requirements
 
 - **Python**: 3.12 or 3.13 (strictly required)
-- **Llama Stack Server**: With at least one foundation model, one embedding model, and vector database configured
 - **Operating System**: macOS or Linux
+- **(Optional) Llama Stack Server**: With at least one foundation model, one embedding model, and vector database configured
+
+
+!!! note "External models and vector database integration"
+    `ai4rag` is designed to be provider-agnostic.
+    It means you can use any model from any source as long as it satisfies `BaseFoundationModel` interface.
+    The same rule applies to embedding model.
+    Custom vector database cannot be explicitly passed to the experiment configuration at this moment, but it can be handled by working the project and delivering custom `VectorStore` implementation.
 
 ---
 
@@ -13,10 +20,12 @@
 Install ai4rag using pip:
 
 ```bash
-pip install ai4rag
+pip install "git+https://github.com/IBM/ai4rag.git@main"
 ```
 
 This installs the core package with all required dependencies.
+Using `"@main"` will download and install latest version of `ai4rag`.
+If you want to use specific version, please use e.g. `"@v0.1.1"`
 
 ---
 
@@ -30,6 +39,8 @@ git clone https://github.com/IBM/ai4rag.git
 cd ai4rag
 
 # Install in editable mode with dev dependencies
+python -m venv venv
+source venv/bin/activate
 pip install -e ".[dev]"
 ```
 
@@ -40,41 +51,15 @@ The `dev` optional dependencies include:
 - Documentation tools (`mkdocs`, `mkdocs-material`)
 - Development utilities (`beautifulsoup4`, `pypdf`, `dotenv`)
 
----
 
-## Optional Dependencies
-
-You can install specific optional dependency groups:
-
-### Testing Only
-
-```bash
-pip install -e ".[test]"
-```
-
-Includes: `pytest`, `pytest-cov`, `pytest-mock`, `psutil`, `nbformat`
-
-### Code Quality Only
-
-```bash
-pip install -e ".[code_check]"
-```
-
-Includes: `pylint`, `black`
-
-### Documentation Only
-
-```bash
-pip install -e ".[docs]"
-```
-
-Includes: `mkdocs`, `mkdocs-material`, `mkdocstrings`, and related plugins
+To see what specific requirement groups are available, please look at the `pyproject.toml` file in the project's root folder.
 
 ---
 
 ## Llama Stack Setup
 
-ai4rag requires a running Llama Stack server. Follow these steps:
+`ai4rag` can be used with Llama Stack server as the foundation models, embedding models and vector database provider.
+Follow these steps:
 
 ### 1. Install Llama Stack
 
@@ -88,7 +73,7 @@ Create a Llama Stack configuration with:
 
 - At least one **foundation model** (e.g., `ollama/llama3.2:3b`)
 - At least one **embedding model** (e.g., `ollama/nomic-embed-text:latest`)
-- A **vector database** (e.g., Milvus or ChromaDB)
+- A **vector database** (e.g., Milvus lite or ChromaDB)
 
 Refer to the [Llama Stack documentation](https://llamastack.github.io/docs/) for detailed setup instructions.
 
@@ -98,7 +83,7 @@ Refer to the [Llama Stack documentation](https://llamastack.github.io/docs/) for
 llama-stack run <your-config.yaml>
 ```
 
-Note the server URL and API key for use in ai4rag.
+Note the server URL and API key for use in `ai4rag`.
 
 ---
 
@@ -108,8 +93,8 @@ Store your Llama Stack credentials securely in a `.env` file:
 
 ```bash
 # .env
-BASE_URL=http://localhost:8000
-APIKEY=your-api-key-here
+BASE_URL="<llama_stack_server_url>"
+API_KEY="<llama_stack_server_api_key>"
 ```
 
 !!! warning "Security"
@@ -119,19 +104,19 @@ Load environment variables in your code:
 
 ```python
 import os
-from dotenv import load_dotenv
+from dotenv import load_dotenv, find_dotenv
 
-load_dotenv()
+load_dotenv(find_dotenv())
 
 base_url = os.getenv("BASE_URL")
-api_key = os.getenv("APIKEY")
+api_key = os.getenv("API_KEY")
 ```
 
 ---
 
 ## Verify Installation
 
-Check that ai4rag is installed correctly:
+Check that `ai4rag` is installed correctly:
 
 ```python
 import ai4rag
@@ -151,7 +136,7 @@ client = LlamaStackClient(
 
 # List available models
 models = client.models.list()
-print(f"Available models: {[m.identifier for m in models]}")
+print(f"Available models: {[m.id for m in models]}")
 ```
 
 ---
@@ -159,40 +144,6 @@ print(f"Available models: {[m.identifier for m in models]}")
 ## Next Steps
 
 - [Quick Start Guide](quick-start.md) - Run your first optimization
-- [Configuration](configuration.md) - Detailed configuration options
 - [User Guide](../user-guide/overview.md) - Comprehensive usage documentation
 
 ---
-
-## Troubleshooting
-
-### Python Version Issues
-
-If you encounter version conflicts:
-
-```bash
-# Check your Python version
-python --version
-
-# Use a specific Python version
-python3.13 -m pip install ai4rag
-```
-
-### Llama Stack Connection Errors
-
-- Verify the server is running: `curl http://localhost:8000/health`
-- Check your `BASE_URL` and `APIKEY` in `.env`
-- Review Llama Stack logs for errors
-
-### Dependency Conflicts
-
-If you experience dependency conflicts:
-
-```bash
-# Create a fresh virtual environment
-python -m venv ai4rag-env
-source ai4rag-env/bin/activate  # On Windows: ai4rag-env\Scripts\activate
-
-# Install ai4rag
-pip install ai4rag
-```
